requirements.txt
speechrecognition
torch
whisper
pyserial
freetype-py
